Reworked GPIO management to save pin related info longer before supplying them to the driver. Also allows the driver wrappers to specify the type of pin that is needed.

+++ esp32s3_hw/config.rs
@@ -6,13 +6,13 @@ pub const LEDS_MAX_X: usize = 8;
 pub const LEDS_MAX_Y: usize = 32;
 
 // Led matrix gpio + channel
-pub const LEDS_IN: usize = 18;
-pub const LEDS_CHANNEL: usize = 3;
+pub const LEDS_IN: u8 = 18;
+pub const LEDS_CHANNEL: u8 = 3;
 
 // Audio input gpio + i2s
-pub const AUDIO_IN_BCLK: usize = 5;
-pub const AUDIO_IN_DIN: usize = 7;
-pub const AUDIO_IN_WS: usize = 6;
-pub const AUDIO_IN_I2S: usize = 0;
+pub const AUDIO_IN_BCLK: u8 = 5;
+pub const AUDIO_IN_DIN: u8 = 7;
+pub const AUDIO_IN_WS: u8 = 6;
+pub const AUDIO_IN_I2S: u8 = 0;
 
 pub const AUDIO_SAMPLE_RATE: u32 = 48000;
